requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
gunicorn==21.2.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
streamlit==1.31.0 